refactor: 認証関連の処理をAuthControllerに移動
